requirements.txt
albumentations==1.3.1
imageio==2.31.3
lmdb==1.4.1
numpy==1.23.5
numpy==1.24.1
Pillow==9.4.0
Pillow==9.3.0
Pillow==10.1.0
piq==0.8.0
pytorch_msssim==0.2.1
scipy==1.11.4
tqdm==4.66.1

